Low: fencer: check "on" support when checking whether device can fence
An initiator calls can_fence_host_with_device() to create a list of devices
that can gence a given target. If the requested action is "on", and a device
doesn't support "on", return false immediately rather than wait until the fence
attempt to determine it can't be done.

This allows for better log messages and slightly faster response.

Fixes T511

+++ daemons/fenced/fenced_commands.c
@@ -2104,6 +2104,13 @@ can_fence_host_with_device(stonith_device_t *dev,
         goto search_report_results;
     }
 
+    // Short-circuit if device does not support action
+    if (pcmk__str_eq(action, "on", pcmk__str_none)
+        && !pcmk_is_set(dev->flags, st_device_supports_on)) {
+        check_type = "Agent does not support 'on'";
+        goto search_report_results;
+    }
+
     /* Short-circuit query if this host is not allowed to perform the action */
     if (pcmk__str_eq(action, "reboot", pcmk__str_casei)) {
         /* A "reboot" *might* get remapped to "off" then "on", so short-circuit
